Warn when running Cookiecutter with incompatible python version
